Rename CardPage to CardsPage in all files and functions
