+++ client/src/KnowNativePage/CardsPage/CardsPage.scss
